+++ templates/Users/sign_in.php
@@ -7,16 +7,18 @@ $this->set('bodyClasses', 'login');
 <div id="classicLogin" style="display:none" class="loginAlternative">
     <h2>Login</h2>
     <ul>
-        <li>Use your e-mail address and DHCR password to log in.</li>
+        <li>Use your institutional e-mail address and DHCR password to log in.</li>
         <li>Reset your password if you forgot it or never set one.</li>
         <li>If you do not have a DHCR account<?php // and were not able to log in using the federated login 
                                                 ?>,
             please use the <?= $this->Html->link('registration form', '/users/register') ?> to create one.</li>
     </ul>
-    <p class="notice">You need to log in to the DHCR if you are a lecturer or other academic institution member and
-        want to add or curate course metadata. All other audience may use the publicly available content and functionality
-        freely. Watch this <a href="https://www.youtube.com/embed/s-YsnpKCaUE">video tutorial</a> to learn how to browse or
-        filter the courses in the registry.</p>
+    <p class="notice">
+        Registration and login details are required only for those members who actively contribute to the DH Course registry
+        (e.g. enter and curate course metadata). All other users are free to use the publicly available content and features
+        without any restrictions. Watch this <a href="https://www.youtube.com/embed/s-YsnpKCaUE">video tutorial</a> to learn
+        how to browse or filter the courses in the registry.
+    </p>
     <div class="users form">
         <?= $this->Form->create() ?>
         <?= $this->Form->control('email', ['label' => 'Email Address', 'autofocus' => 'true']) ?>