+++ RELEASE_NOTES.md
@@ -1,7 +1,37 @@
 # Release notes
 
+## Version 2025-07
+*Release date: 2025-07-12*
+
+### New feature
+- Added fields for course name and description in original language #131
+
+### Small change
+- Updated dependencies
+
+### Bug fix
+- Fix: Limit "Avg external resources per course" on 1 decimal in summary statistics #48
+
+
+## Version 2025-06
+*Release date: 2025-07-12*
+
+### Small changes
+- Added total numer of results to main page #77
+- Updated text on registration page #123
+- Updated text on login page #123
+- Make unpublished courses visible for national moderators #125
+- Added external resources overview page in login area #48
+- Added external resource counts to summary statistics dashboard #48
+- Fix: Updated field is not filled with a value #48
+- Fixed deprecation warning in GenSearchListCommand
+- Fixed deprecation warning in GenSitemapCommand
+- Fixed deprecation warning in ReviewRemindersCommand
+- Updated dependencies
+
+
 ## Version 2025-05/2
-*Release date: 2025-06-04* 
+*Release date: 2025-06-04*
 
 ### Small changes
 - Improved visualization on course detail page, in the case that one of the optional external resource fields (Label, Type, Affiliation) are empty. #48
@@ -17,7 +47,7 @@
 *Release date: 2025-05-14*
 
 ### New feature
-- Add additional metadata for registering RI datasets / training / learning material used in a course or programme #48
+- Add additional metadata for registering RI datasets / training / learning material used in a course or programme (External Resources) #48
 
 ### Changes
 - Added documentation for application installation in local development environment #33